requirements.txt
pytest
allure-pytest
selene>=2.0.b17
selenium==4.5.0